+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Changed
 - Use the `global` coordinate system by default in the remaining readers that were still using the `pixels` coordinate system
+- Use `prior_shapes_key: auto` in all Snakemake config - it will automatically find the right key based on the technology
 
 ### Removed
 - Removed the `open-cv` dependency (#239)